+++ frontend/javascript/jatek/entitások/jatekos.js
@@ -1,0 +1,1 @@
+console.log("Játékos.js betöltve");
