+++ frontend/javascript/jatek/kaboomBetolto.js
@@ -1,0 +1,1 @@
+console.log("Kaboom betöltve");
